Fix assimp build on uncommon architectures
This adapts the patch from upstream r1079
http://assimp.svn.sourceforge.net/viewvc/assimp?view=revision&revision=1079

The build should not error out when it encounters an architecture that
has not been explicitly included. The truly annoying part of the error
was that it *only* affected how a debug string was constructed.

Using the byte order as provided by the build system is much nicer and
hopefully cleaner.